refactor(files): extract sidebar SearchResults into own module
