+++ src/components/files/sidebar.tsx
@@ -1,9 +1,10 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { FolderOpen, Search, X } from "lucide-react";
 import { Button, Icon } from "@/components/primitives";
-import { basename, dirname, shortcutLabel, startWindowDrag, walkMarkdownFiles, type FileEntry, type FlatFileEntry } from "@/lib";
+import { basename, shortcutLabel, startWindowDrag, type FileEntry } from "@/lib";
 import emptyTowerUrl from "@/assets/mascot/empty-m.png";
 import { FileTree, type NewEntry } from "./file-tree";
+import { SearchResults } from "./sidebar-search";
 
 type SidebarProps = {
   open: boolean;
@@ -274,74 +275,3 @@ export function Sidebar({
   );
 }
 
-type SearchResultsProps = {
-  rootPath: string;
-  query: string;
-  activePath: string | null;
-  treeVersion?: number;
-  onSelect: (path: string) => void;
-};
-
-const MAX_RESULTS = 80;
-
-function SearchResults({ rootPath, query, activePath, treeVersion = 0, onSelect }: SearchResultsProps) {
-  const [index, setIndex] = useState<FlatFileEntry[] | null>(null);
-
-  useEffect(() => {
-    let cancelled = false;
-    setIndex(null);
-    walkMarkdownFiles(rootPath)
-      .then((items) => {
-        if (!cancelled) setIndex(items);
-      })
-      .catch(() => {
-        if (!cancelled) setIndex([]);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [rootPath, treeVersion]);
-
-  const results = useMemo(() => {
-    if (!index) return null;
-    const q = query.trim().toLowerCase();
-    if (!q) return [];
-    const out: FlatFileEntry[] = [];
-    for (const e of index) {
-      if (e.rel.toLowerCase().includes(q)) {
-        out.push(e);
-        if (out.length >= MAX_RESULTS) break;
-      }
-    }
-    return out;
-  }, [index, query]);
-
-  if (!index) return <div className="mdv-tree__loading">indexing…</div>;
-  if (results && results.length === 0) {
-    return <div className="mdv-tree__empty">no matches</div>;
-  }
-
-  return (
-    <ul className="mdv-search-results" role="listbox">
-      {results?.map((r) => {
-        const dir = dirname(r.rel);
-        const isActive = r.path === activePath;
-        return (
-          <li key={r.path}>
-            <button
-              type="button"
-              className={`mdv-search-result${isActive ? " is-active" : ""}`}
-              title={r.path}
-              onClick={() => onSelect(r.path)}
-            >
-              <span className="mdv-search-result__name">{r.name}</span>
-              {dir && dir !== "/" ? (
-                <span className="mdv-search-result__dir">{dir}</span>
-              ) : null}
-            </button>
-          </li>
-        );
-      })}
-    </ul>
-  );
-}
